Compiling and deleting files not necessary for the project. Multiple of the files were structured around the old COINr (v6)-package that has since been updated.
The main files still stand and have been updated to new COINr and code has been cleaned.

Follow-up is to finalise the policy database analysis for NOURISHING database, finish visualisation files for MOVING and NOURISHING. I may split the new COINr methodology in two to make it easier to click-and-run.